chore(docs): Update `Mavrick's` commit guidelines
Added more details to the commit guidelines as they needed to be more
clear

Removed the mettion of how to write commit messges in the `merge` section
as it would have been seen as a double up of information.

Restuecterd the ordering to make more senes when reading the guidelines.