test: Add testing methods to allow testing of sqlite
I have added a some tests that allow for use to verify that the sqlite
query are correct by using `:memory:` in the jdbc connection string
as well as setting testing mode on the SqliteConnection class.

this allows for in memory sqlite to run the real query with no need for
mock database.

I am indeed the goat